% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalHerfindahl2.R
\name{generalHerfindahl}
\alias{generalHerfindahl}
\title{Calculate the Herfindahl index and competition for non-MEDSL data}
\usage{
generalHerfindahl(data, office_field, district_field, vote_field,
  cand_field, other_fields)
}
\arguments{
\item{data}{The long dataframe with the election results of interest}

\item{office_field}{The name of the office/race field that the results will be summarized by.}

\item{district_field}{The name of the district field that when combined with the office will provide the ID by which to summarise the 
results by}

\item{vote_field}{The name of the field for vote totals.}

\item{cand_field}{The name of the field for candidate, which when combined with the office and other fields, should provide the atomic unit
by which to square the proportions of.}

\item{other_fields}{The name of the other identifying fields (i.e. district, state, county, etc.) by which the user can score levels of 
competition to. Will be included in the group_by command.}
}
\value{
The dataframe with the competition scores by race for the provided dataset, and the following fields: 
    \itemize{
    \item herf_score = The herfindahl index score of votes, where scores of one approximate elections where the vote share went entirely to
    one candidate, and scores approaching zero fragmented between numerous candidates
    \item max_vote = The vote total for the candidate with the most votes for a given election, i.e. the winning total.
    \item max_cand = The candidate with the highest vote total for a given election, i.e. the winner. 
    \item competition = The reverse herfindahl index such that higher scores equate to greater competition, and scores closer to zero
    less competition. 
    
}
}
\description{
This function takes the long datasets for other election information and computes the Herfindahl index measure of competition so that 
one can compare it to the raw precinct results
}
\examples{
house_all <- read.csv("1976-2018-house.csv")
test_house<-generalHerfindahl(house_all, office_field = "office", district_field = "district", vote_field = 'candidatevotes', 
cand_field = "candidate", other_fields = c("state_fips", "year", "stage") )

}
