% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medslCompetition.R
\name{medslCompetition}
\alias{medslCompetition}
\title{Calculates the competition for MEDSL non-local elections data}
\usage{
medslCompetition(data, dataverse, mag = c(TRUE, FALSE))
}
\arguments{
\item{data}{The data frame precinct object with the local election data. Should be formatted with proper MEDSL standards.}

\item{dataverse}{The option to provide a dataverse to subset by.}

\item{mag}{The optional command that if true, will trigger a revised herfindahl index in order to deal with multimemeber districts}
}
\value{
The the competition dataframe object. Will have the following fields:
   \itemize{
   \item office = The name of the office for the given race.
   \item district = the district for the given race. For counties and jurisdictions within MEDSL data with empty districts, coded as the 
   name of the county and/or jurisdiction. Additionally, in the event that there is a numeric value, the name of the appropriate county
   or jurisdiction is pasted on
   \item stage = The stage for a given race, i.e. gen, primary, runoff, etc. 
   \item special = Whether an election was a sepcial election. 
   \item herf_score = The sum of squared proportions for all of the candidates within the given precinct data. Scores closer to one reflect
   complete lack of diversity in the vote share, centered around a single candidate, and scores closer to zero reflect complete 
   fragmentation and diversity within the voteshare. Note that scores closer to zero might be due to higher district magnitudes, or in the
   event of improperly cleaned data, errors in the fields of office, county, jurisdiction, candidate, and/or party.  
   \item state_fips = The FIPs code for a state. 
   \item partisan = The race had at least one of the two major parties openly contest the race. 1 if partisan, and 0 otherwise. 
   \item county_count = The number of counties for a given race. 
   \item jurisdiction_count - The number of jurisdictions within a given race.
   \item party_count = The number of unique parties found to have contested elections. Note that this also includes blanks. Note2: 
   Given the writing of this package on January 23, 2020, the MEDSL data has yet to have the party_simplified category coded in. 
   \item total_vote = The total votes cast for a given race. 
   \item max_vote = The maximum vote share for the candidate within the race, and usually the sole winner. 
   \item max_cand = The name of the candidate/ballot item with the maximum vote share. 
   \item cand_count = The number of unique candidates found within the data. Note that in the event of detailed write-ins for states 
   such as Vermont and New Hampshire, these might number in the hundreds. 
   \item competition = The reverse of the herfindahl score such that scores closer to one equate to greater competition, and scores of 
   zero to non-competition. 
   \item total_vote2 = The sum of votes used to calculate the herfindahl index when there are multimember districts. Only calculated if 
   mag == TRUE
   \item magnitude = The district's magnitude 
   
}
}
\description{
Calculates the competition for MEDSL non-local elections data
}
\examples{
mn_results <- read.csv("2018-il-local_precinct.csv")
mn_comp <- medslCompetition(mn_results)
head(mn_results$magnitude)
1 1 3 1 3 
mn_comp2 <- localCompetition(mn_results,mag=TRUE)


}
